Splits Venue into components
